rename enroll to capi enroll-to-bo
