+++ cmd/crowdsec-cli/capi.go
@@ -166,5 +166,70 @@ func NewCapiCmd() *cobra.Command {
 		},
 	}
 	cmdCapi.AddCommand(cmdCapiStatus)
+
+	cmdEnroll := &cobra.Command{
+		Use:   "enroll-to-bo [enroll-key]",
+		Short: "Enroll this instance to https://app.crowdsec.net [requires local API]",
+		Long: `
+Enroll this instance to https://app.crowdsec.net
+		
+You can get your enrollment key by creating an account on https://app.crowdsec.net.
+After running this command your will need to validate the enrollment in the webapp.`,
+		Example: "cscli enroll-to-bo YOUR-ENROLL-KEY",
+		Args:    cobra.ExactArgs(1),
+		PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
+			if err := csConfig.LoadAPIServer(); err != nil || csConfig.DisableAPI {
+				log.Fatal("Local API is disabled, please run this command on the local API machine")
+			}
+			if csConfig.API.Server.OnlineClient == nil {
+				log.Fatalf("no configuration for crowdsec API in '%s'", *csConfig.FilePath)
+			}
+			if csConfig.API.Server.OnlineClient.Credentials == nil {
+				log.Fatal("You must configure CAPI with `cscli capi register` before enrolling your instance")
+			}
+			return nil
+		},
+		Run: func(cmd *cobra.Command, args []string) {
+			password := strfmt.Password(csConfig.API.Server.OnlineClient.Credentials.Password)
+			apiURL, err := url.Parse(csConfig.API.Server.OnlineClient.Credentials.URL)
+			if err != nil {
+				log.Fatalf("Could not parse CAPI URL : %s", err)
+			}
+
+			if err := csConfig.LoadHub(); err != nil {
+				log.Fatalf(err.Error())
+			}
+
+			if err := cwhub.GetHubIdx(csConfig.Hub); err != nil {
+				log.Fatalf("Failed to load hub index : %s", err)
+				log.Infoln("Run 'sudo cscli hub update' to get the hub index")
+			}
+
+			scenarios, err := cwhub.GetUpstreamInstalledScenariosAsString()
+			if err != nil {
+				log.Fatalf("failed to get scenarios : %s", err.Error())
+			}
+
+			if len(scenarios) == 0 {
+				scenarios = make([]string, 0)
+			}
+
+			c, _ := apiclient.NewClient(&apiclient.Config{
+				MachineID:     csConfig.API.Server.OnlineClient.Credentials.Login,
+				Password:      password,
+				Scenarios:     scenarios,
+				UserAgent:     fmt.Sprintf("crowdsec/%s", cwversion.VersionStr()),
+				URL:           apiURL,
+				VersionPrefix: "v2",
+			})
+			_, err = c.Auth.EnrollWatcher(context.Background(), args[0])
+			if err != nil {
+				log.Fatalf("Could not enroll instance: %s", err)
+			}
+			log.Infof("Watcher successfully enrolled. Visit https://app.crowdsec.net to accept it.")
+		},
+	}
+
+	cmdCapi.AddCommand(cmdEnroll)
 	return cmdCapi
 }
